requirements.txt
vllm>=0.6.0
transformers>=4.45.0
datasets>=3.0.0
huggingface_hub>=0.25.0
torch>=2.4.0
peft>=0.13.0
trl>=0.12.0
accelerate>=1.0.0
pyyaml>=6.0
matplotlib>=3.8.0
pandas>=2.1.0
tqdm>=4.66.0
pylint>=3.0.0
radon>=6.0.0
numpy>=1.26.0
scipy>=1.11.0
seaborn>=0.13.0
